test(riviere-extract-ts): split coverage tests into one concept per test
