+++ packages/riviere-extract-ts/use-cases/src/features/extract/commands/command-error-coverage.spec.ts
@@ -8,24 +8,14 @@ import {
 } from '../../../__fixtures__/command-test-fixtures'
 import { AddDomain } from './add-domain'
 import { AddComponent } from './add-component'
-import { AddSource } from './add-source'
 import { DefineCustomType } from './define-custom-type'
 import { DefineRelationshipType } from './define-relationship-type'
-import { EnrichComponent } from './enrich-component'
 import { FinalizeGraph } from './finalize-graph'
 import { InitGraph } from './init-graph'
 import { LinkComponents } from './link-components'
 import { LinkExternal } from './link-external'
 import { LinkHttp } from './link-http'
-import { ValidateGraph } from './validate-graph'
 import { RiviereProjectRepository } from '../data-access/riviere-project/riviere-project-repository'
-
-class UnexpectedBuilderFailure extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = 'UnexpectedBuilderFailure'
-  }
-}
 
 function createProject(): RiviereProject {
   return RiviereProject.start({
@@ -34,25 +24,6 @@ function createProject(): RiviereProject {
       sources: [{ repository: 'https://github.com/org/repo' }],
     },
   }).data
-}
-
-function createProjectWithApi(): RiviereProject {
-  const project = createProject()
-  project.amendGraph((builder) =>
-    builder.addApi({
-      apiType: 'REST',
-      domain: 'orders',
-      httpMethod: 'POST',
-      module: 'core',
-      name: 'CreateOrder',
-      path: '/orders',
-      sourceLocation: {
-        repository: 'https://github.com/org/repo',
-        filePath: 'src/create-order.ts',
-      },
-    }),
-  )
-  return project
 }
 
 describe('command error path coverage', () => {
@@ -113,7 +84,7 @@ describe('command error path coverage', () => {
     expect(result.result).toMatchObject({ code: 'VALIDATION_ERROR', success: false })
   })
 
-  it('returns validation errors from define-custom-type for invalid property types', () => {
+  it('returns validation error from define-custom-type for invalid required property type', () => {
     expect(
       new DefineCustomType(new RiviereProjectRepository()).execute({
         description: undefined,
@@ -123,6 +94,9 @@ describe('command error path coverage', () => {
         requiredProperties: { retries: { type: 'not-a-type' } },
       }).result,
     ).toMatchObject({ code: 'VALIDATION_ERROR', success: false })
+  })
+
+  it('returns validation error from define-custom-type for invalid optional property type', () => {
     expect(
       new DefineCustomType(new RiviereProjectRepository()).execute({
         description: undefined,
@@ -200,7 +174,7 @@ describe('command error path coverage', () => {
     ).toMatchObject({ code: 'GRAPH_NOT_FOUND', success: false })
   })
 
-  it('returns component not found from link-external and a validation error from link-components', () => {
+  it('returns component not found from link-external', () => {
     const project = createProject()
     vi.spyOn(RiviereProjectRepository.prototype, 'load').mockReturnValue(project)
     expect(
@@ -213,6 +187,11 @@ describe('command error path coverage', () => {
         type: undefined,
       }).result,
     ).toMatchObject({ code: 'COMPONENT_NOT_FOUND', success: false })
+  })
+
+  it('returns validation error from link-components for invalid input', () => {
+    const project = createProject()
+    vi.spyOn(RiviereProjectRepository.prototype, 'load').mockReturnValue(project)
     expect(
       new LinkComponents(new RiviereProjectRepository()).execute({
         from: 'not-a-valid-id',
@@ -226,7 +205,7 @@ describe('command error path coverage', () => {
     ).toMatchObject({ code: 'VALIDATION_ERROR', success: false })
   })
 
-  it('returns a graph corrupted error from add-domain and link-external', async () => {
+  it('returns graph corrupted error from add-domain', async () => {
     const { mkdir, writeFile } = await import('node:fs/promises')
     await mkdir(join(ctx.testDir, '.riviere'), { recursive: true })
     await writeFile(graphLocation(), '{invalid', 'utf-8')
@@ -238,6 +217,12 @@ describe('command error path coverage', () => {
         systemType: 'domain',
       }).result,
     ).toMatchObject({ code: 'GRAPH_CORRUPTED', success: false })
+  })
+
+  it('returns graph corrupted error from link-external', async () => {
+    const { mkdir, writeFile } = await import('node:fs/promises')
+    await mkdir(join(ctx.testDir, '.riviere'), { recursive: true })
+    await writeFile(graphLocation(), '{invalid', 'utf-8')
     expect(
       new LinkExternal(new RiviereProjectRepository()).execute({
         from: 'orders:core:api:source',
@@ -250,102 +235,7 @@ describe('command error path coverage', () => {
     ).toMatchObject({ code: 'GRAPH_CORRUPTED', success: false })
   })
 
-  it('rethrows unexpected errors from add-source, add-domain, validate-graph, and finalize-graph', () => {
-    const project = createProject()
-    vi.spyOn(RiviereProjectRepository.prototype, 'load').mockReturnValue(project)
-    vi.spyOn(project, 'amendGraph').mockImplementation(() => {
-      throw new UnexpectedBuilderFailure('boom')
-    })
-    expect(() =>
-      new AddSource(new RiviereProjectRepository()).execute({
-        graphFileLocation: graphLocation(),
-        repository: 'https://github.com/org/x',
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-    expect(() =>
-      new AddDomain(new RiviereProjectRepository()).execute({
-        description: 'x',
-        graphFileLocation: graphLocation(),
-        name: 'payments',
-        systemType: 'domain',
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-    expect(() =>
-      new ValidateGraph(new RiviereProjectRepository()).execute({
-        graphFileLocation: graphLocation(),
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-    expect(() =>
-      new FinalizeGraph(new RiviereProjectRepository()).execute({
-        graphFileLocation: graphLocation(),
-        outputPath: '/out',
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-  })
-
-  it('rethrows unexpected errors from enrich, link-components, and link-external', () => {
-    const project = createProject()
-    vi.spyOn(RiviereProjectRepository.prototype, 'load').mockReturnValue(project)
-    vi.spyOn(project, 'amendGraph').mockImplementation(() => {
-      throw new UnexpectedBuilderFailure('boom')
-    })
-    expect(() =>
-      new EnrichComponent(new RiviereProjectRepository()).execute({
-        businessRules: [],
-        entity: undefined,
-        emits: [],
-        graphFileLocation: graphLocation(),
-        id: 'orders:core:domainop:place-order',
-        modifies: [],
-        reads: [],
-        signature: undefined,
-        stateChanges: [],
-        validates: [],
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-    expect(() =>
-      new LinkComponents(new RiviereProjectRepository()).execute({
-        from: 'orders:core:api:source',
-        graphFileLocation: graphLocation(),
-        targetDomain: 'orders',
-        targetModule: 'core',
-        targetName: 'Place Order',
-        targetType: 'UseCase',
-        type: undefined,
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-    expect(() =>
-      new LinkExternal(new RiviereProjectRepository()).execute({
-        from: 'orders:core:api:source',
-        graphFileLocation: graphLocation(),
-        targetDomain: undefined,
-        targetName: 'Stripe',
-        targetUrl: undefined,
-        type: undefined,
-      }),
-    ).toThrow(UnexpectedBuilderFailure)
-  })
-
-  it('rethrows unexpected errors from link-http', () => {
-    const project = createProjectWithApi()
-    vi.spyOn(RiviereProjectRepository.prototype, 'load').mockReturnValue(project)
-    vi.spyOn(project, 'amendGraph').mockImplementation(() => {
-      throw new UnexpectedBuilderFailure('link-http boom')
-    })
-    expect(() =>
-      new LinkHttp(new RiviereProjectRepository()).execute({
-        graphFileLocation: graphLocation(),
-        httpMethod: 'POST',
-        linkType: undefined,
-        path: '/orders',
-        targetDomain: 'orders',
-        targetModule: 'core',
-        targetName: 'Place Order',
-        targetType: 'UseCase',
-      }),
-    ).toThrow('link-http boom')
-  })
-  it('returns graph not found and a validation error from add-component', async () => {
+  it('returns graph not found from add-component', () => {
     const base = {
       componentType: 'UseCase',
       domain: 'orders',
@@ -359,6 +249,18 @@ describe('command error path coverage', () => {
       code: 'GRAPH_NOT_FOUND',
       success: false,
     })
+  })
+
+  it('returns validation error from add-component for a corrupted graph', async () => {
+    const base = {
+      componentType: 'UseCase',
+      domain: 'orders',
+      filePath: 'f',
+      graphFileLocation: graphLocation(),
+      module: 'core',
+      name: 'x',
+      repository: 'r',
+    }
     const { mkdir, writeFile } = await import('node:fs/promises')
     await mkdir(join(ctx.testDir, '.riviere'), { recursive: true })
     await writeFile(graphLocation(), '{invalid', 'utf-8')
@@ -368,7 +270,7 @@ describe('command error path coverage', () => {
     })
   })
 
-  it('returns validation errors from link-external for invalid input', () => {
+  it('returns validation error from link-external for an invalid input id', () => {
     expect(
       new LinkExternal(new RiviereProjectRepository()).execute({
         from: 'not-an-id',
@@ -379,6 +281,9 @@ describe('command error path coverage', () => {
         type: undefined,
       }).result,
     ).toMatchObject({ code: 'VALIDATION_ERROR', success: false })
+  })
+
+  it('returns validation error from link-external for an invalid type', () => {
     expect(
       new LinkExternal(new RiviereProjectRepository()).execute({
         from: 'orders:core:api:source',
@@ -403,19 +308,5 @@ describe('command error path coverage', () => {
         systemType: 'domain',
       }).result,
     ).toMatchObject({ code: 'GRAPH_CORRUPTED', success: false })
-  })
-
-  it('rethrows unexpected load errors from init-graph', () => {
-    vi.spyOn(RiviereProjectRepository.prototype, 'load').mockImplementation(() => {
-      throw new UnexpectedBuilderFailure('init boom')
-    })
-    expect(() =>
-      new InitGraph(new RiviereProjectRepository()).execute({
-        domains: [{ description: 'Orders', name: 'orders', systemType: 'domain' }],
-        graphFileLocation: graphLocation(),
-        name: 'combined',
-        sources: ['https://github.com/org/repo'],
-      }),
-    ).toThrow('init boom')
   })
 })
